+++ src/components/skeleton/skeleton-interface.ts
@@ -1,0 +1,1 @@
+export type WcsSkeletonAnimation = 'none' | 'glide' | 'pulse';