+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - **react**: create react bindings for WCS components
 - **doc**: deploy documentation on Azure static web app (https://wcs.dev.sncf)
+- **skeleton**: add new component `skeleton-text`
 
 ### Changed